Moved `CStrBuf`, `NegativeI32` to new `utils` mod
They're not TBL-specific, and this move reflects that.
